requirements.txt
pandas>=2.0.0
scikit-learn==1.8.0
joblib>=1.3.2
gradio==4.44.1
huggingface-hub==0.24.0
plotly
imbalanced-learn>=0.12.0
langgraph
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
chromadb
python-dotenv